PDS-792 - Harvest: Handle Search Core config fields that need to go to a Registry to get values  (#67)
* PDS-792 - Harvest: Update Search Core configurations.

This updates how the tool populates Search Core fields that used to go
out to a Registry to get its values. They were denoted by a `.` period
character (i.e. instrument_ref.instrument_name). For instrument_name and
instrument_host_name, Harvest Policy config was updated to extract that
information from the label itself. For resources, it was decided to use
a lookup table to populate these fields. A script was created to get the
resources currently registered in Search. It can be run like so:

`%> python get_resources.py`

Output is a JSON-formatted file that contains the resource LID, the
resource_name and resource_url. This file is read in by Harvest at run
time.