Registry and Harvest Hot Fixes merged from master (#89)
* PDS-796 Changed 'clean' mode to 'uninstall' (#79)

* PDS-822 - Harvest: Updated to register XPath Solr documents. (#80)

This update will post Solr documents containing a set of XPaths of the
entire label to a Solr collection named `xpath`. This will only occur if
the label blob was successfully registered.

* PDS-822: Update Search Service deployment with xpath collection (#81)

* PDS-822: Update Search Service deployment with xpath collection

updates includes modifications to the deploy script and the Dockerfile to
create a new collection for the xpath data.

Also includes some cleanup and streamlining of the installer script.

* Cleanup merge conflicts

* Cleanup script removing output comments

* PDS-804: Update Harvest and Search to for lid and lidvid fields (#82)

* PDS-804 - Harvest: Update config with lids

Previous Search Core config translated lid to identifier, but that is the wrong place to
do this. Instead, we will now ingest `lid` directly into the Search Service and update
the Solr config and schema to copy to the identifier field for the API.

* PDS-804 - Search: Update schema and config for lid/lidvids

* Update schema to copy lidvid and lid to identifier field for API
* Update solrconfig with new generateLIDVID updateProcessorChain to concatenate lid and version_id

* Update harvest example with inline comments (#83)

* Search: Lower heap space to 2GiB (#84)

Depending on your Docker/system RAM and physical memory, anything much larger than these
seems to crash Solr. This may be a safer bet.

* Update lidvid and identifier to be multivalued

* Fix typo with identifier field declaration

* PDS-831 - Harvest: Update how the tool versions products (#87)

* PDS-831 - Harvest: Update how the tool versions products

- Add the version id to the blobName when ingesting to the .system
collection.

- Set the `id` as the LIDVID when posting documents to the xpath
collection.

* PDS-831: Update solrconfig to use lidvid as uniqueKey

* PDS-849 / PDS-850: Update Registry for displaying latest product version (#88)

* PDS-849: Updated Registry search requestHandler for latest product

Solr provides the ability to perform a "group by" in a query:
https://lucene.apache.org/solr/guide/6_6/result-grouping.html

Using this as a template, we added various parameters to the default
query for the search/ and archive-filter/ request handlers, including:

* group = true - enable the grouping functionality
* group.field = lid - we have multiple versions of products based upon lids
* ground.sort = version_id desc - order groupings by version_id in descending order.
* group.format = simple - simplifies the format a bit for parsing
* group.ngroups = true - enable including number of groups in results. critical to getting a valid number of actual search results (not including older versions)
* group.truncate = true - enable this so the facets use the ngroups value not the actual number of products
* group.limit = 1 - default value we actually don't specify so it only shows the first value of the group

* PDS-850: Update XSLT to handle new Solr grouping

* Revert "Merge branch 'dev' into master"

This reverts commit 21f92d636869b5aaaa66714a3926b00eb82967f2, reversing
changes made to ec9d6de68f5fef6ce189ec98a76c37bed1ae0f72.