updates to match cognito
Cognito needs three values to be consistent. To do this, modified the cognitoCliendId tag to be an extension of the normalizedString then added the two URLs that are needed. In theory, the two URLs do not need to change for an ops service so have them defaulted to something right now and left them as optional. Can and should change these when the real multitenacy goes live.

Moved the self signed certificate trust to the server URL since that is whwere it belongs - the registry is not the problem especially with cognito but a standalone server for testing or development can be and probably is.