Update Harvest to support multi-tenancy and OpenSearch Serverless (#152)
* updates to match new common API

The changes remove the local understanding of data and push it to a facade with a series of impls and wrappers to hide the actual understanding of opensearch in registry-common.

* move to the connection URL

This pulls all of the understanding of connecting to opensearch out of harvest and into registry-common. It uses the same URL style as the command line registry-mgr.

* update all of the examples

These now reflect the new harvest configuration as well as the connections. Use xmllint to verify that harvest configuration is valid.

They point at the schema here at github for now.